agents/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "hypothesis>=6.151.9",
    "langgraph>=1.0.9",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "openrouter>=0.6.0",
    "pandas>=3.0.1",
    "playwright>=1.58.0",
    "pydantic>=2.12.5",
    "pypdf>=6.7.1",
    "python-dotenv>=1.2.1",
    "redis>=7.2.0",
    "seaborn>=0.13.2",
]
